1. a little off topic but implemented Settings page with my cat showing that it's still in development :) 2. Fixed header to display date time autonomously 3. Renamed payments to invites (more logical) 4. Cleaned up previous header date time logic
