+++ frontend/src/app/layout/content/nav-menu/nav-menu.html
@@ -18,7 +18,7 @@
           <a href="#" (click)="$event.preventDefault(); goCreatePmatch() ; close()" class="hover:text-blue-600 flex font-bold items-center gap-3">
             ⭐ {{ isAdmin ? 'Créer Match Public' : 'Créer Match Privé' }}
           </a>
-          <a routerLink="#" (click)="close()" class="hover:text-blue-600 flex font-bold items-center gap-3">🛠️ REGLAGES</a>
+          <a [routerLink]="currentMatricule ? ['/home', currentMatricule, 'settings'] : ['/home']" (click)="close()" class="hover:text-blue-600 flex font-bold items-center gap-3">🛠️ REGLAGES</a>
           <button (click)="onLogout()" class="mt-4 text-left hover:text-red-500 flex font-bold items-center gap-3 p-1">⏻ SE DECONNECTER</button>
         </nav>
       </div>
@@ -35,7 +35,7 @@
         <a href="#" (click)="$event.preventDefault(); goJoinPublic()" class="hover:text-blue-600 flex items-center gap-3">➕ Rejoindre Match Public</a>
       }
       <a href="#" (click)="$event.preventDefault(); goCreatePmatch()" class="hover:text-blue-600 flex items-center gap-3">⭐ {{ isAdmin ? 'Créer Match Public' : 'Créer Match Privé' }}</a>
-      <a routerLink="#" class="hover:text-blue-600 flex items-center gap-3">🛠️ Reglages</a>
+      <a [routerLink]="currentMatricule ? ['/home', currentMatricule, 'settings'] : ['/home']" class="hover:text-blue-600 flex items-center gap-3">🛠️ Reglages</a>
       <button (click)="onLogout()" class="hover:text-red-500 flex items-center gap-3 p-1">⏻ Se Déconnecter</button>
     </nav>
   </aside>
