+++ frontend/src/app/layout/content/pages/my-matches/my-matches.html
@@ -1,5 +1,5 @@
 <div class="relative flex flex-col min-h-screen px-4 md:px-8">
-  <app-home-account-header [todayDate]="null"></app-home-account-header>
+  <app-home-account-header></app-home-account-header>
 
   <div class="flex flex-1 mid:flex-row relative">
     <main class="flex-1 p-6 relative flex flex-col items-stretch z-10">
