+++ frontend/src/app/layout/content/pages/new-priv-match/new-priv-match.html
@@ -1,6 +1,6 @@
-<app-home-account-header [todayDate]="today"></app-home-account-header>
+<app-home-account-header></app-home-account-header>
 
-<div class="flex flex-1 flex-col mid:flex-row relative p-8">
+<div class="flex flex-1 mid:flex-row relative p-8">
   <main class="flex-1 p-6 relative flex flex-col items-stretch z-10">
     <div class="w-full max-w-full mb-6">
       <h1 class="text-2xl font-bold">Nouveau match privé organisé par {{ userName }}</h1>
@@ -15,4 +15,3 @@
   <!-- Page-level NavMenu placed here so desktop aside aligns to the right -->
   <app-nav-menu></app-nav-menu>
 </div>
-
