+++ frontend/src/app/layout/content/pages/new-pub-match/new-pub-match.html
@@ -1,4 +1,4 @@
-<app-home-account-header [todayDate]="today"></app-home-account-header>
+<app-home-account-header></app-home-account-header>
 
 <div class="flex flex-1 flex-col mid:flex-row relative p-8">
   <main class="flex-1 p-6 relative flex flex-col items-stretch z-10">
@@ -15,6 +15,3 @@
   <!-- Page-level NavMenu placed here so desktop aside aligns to the right -->
   <app-nav-menu></app-nav-menu>
 </div>
-
-
-
