+++ frontend/src/app/layout/content/pages/settings/settings.html
@@ -1,0 +1,22 @@
+<div class="relative flex flex-col min-h-screen px-4 md:px-8">
+  <app-home-account-header></app-home-account-header>
+
+  <div class="flex flex-1 mid:flex-row relative">
+    <main class="flex-1 p-6 relative flex flex-col items-center justify-center z-10 mt-5">
+      <div class="flex flex-col items-center justify-center text-center p-6 bg-white/60 backdrop-blur-sm border border-gray-200 rounded-xl shadow-lg">
+        <h2 class="text-2xl font-bold text-gray-800 mb-4">Cette section est en cours de développement.</h2>
+        <div class="max-w-md rounded-2xl overflow-hidden shadow-2xl border-4 border-white">
+          <img src="assets/double-nikiya.jpg" alt="Double Nikiya" class="w-full h-auto object-cover">
+        </div>
+        <br/>
+        <h2 class="text-2xl font-bold text-orange-500 mb-4">Voici une photo du chat du développeur pour nous excuser :)</h2>
+
+        <a routerLink="/home" class="mt-4 bg-blue-600 hover:bg-blue-700 text-white font-bold py-2 px-6 rounded-lg transition-colors shadow-md">
+          Retourner
+        </a>
+      </div>
+    </main>
+
+    <app-nav-menu></app-nav-menu>
+  </div>
+</div>
